Provide a configurable way to decide if the cluster diameter is larger than the threshold.
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/trunk/qtclustering@14086 60cd2ccc-2b05-0410-9566-da273a940bff

+++ src/edu/cornell/med/icb/clustering/Clusterer.java
@@ -39,7 +39,7 @@ public interface Clusterer {
      * @return The list of clusters.
      */
     List<int[]> cluster(final SimilarityDistanceCalculator calculator,
-                        final float qualityThreshold);
+                        final double qualityThreshold);
 
     /**
      * Returns the list of clusters produced by clustering.
